UPDATE
The previous commit had an issue with the on-click menu toggles and URL check.

The URL check has been rectified to ensure all feeds have a defined
and non-empty URL.

The tests on the menu were failing even on correct inputs due to an incorrect class selector.
The selector has been rectified to ".menu-icon-link" and is working fine.

The expect values on the menu test have been corrected as well.

The 'menu-hidden'  class results to false when
the menu visible and vice-versa